feat: display property card by id or all data if id prop passed
