+++ frontend/src/components/organism/PropertyCard.jsx
@@ -1,79 +1,41 @@
-// 3rd party
-import { Card, Image } from "react-bootstrap";
-
 // atomic design
-import { button, spinner } from "../atom";
+import { spinner } from "../atom";
 
 // custom hooks
 import useFetch from "../../hooks/useFetch";
-import useAppContext from "../../hooks/useAppContext";
 
 // styling
 import styles from "../../styles/components/organism/Card.module.css";
+import DisplayPropertyCard from "../molecule/DisplayPropertyCard";
 
 /**
  * Displaying of Property in Card format
  * @returns {JSX.Element}
  * @constructor
  */
-const PropertyCard = () => {
-  // state store
-  const { dispatch } = useAppContext();
+const PropertyCard = ({ id }) => {
+  // fetch property by id if id is not undefined
+  const with_id = useFetch(`properties/${id}/`, id !== undefined);
+  // fetch property data if id is undefined
+  const without_id = useFetch("properties/read/", id === undefined);
 
   // fetch data
-  const { data } = useFetch("properties/read/");
+  const { data } = id === undefined ? without_id : with_id;
 
   return (
     <div className={styles.card}>
       {/*display data when it exists*/}
-      {data
-        ? data.map((items) => (
-            <Card key={items.id} style={{ width: "18rem" }}>
-              <Image
-                src={items.image}
-                width="100%"
-                height="200"
-                objectfit="cover"
-              />
-              <Card.Body>
-                <Card.Title>Post code: {items.area_code}</Card.Title>
-                <hr />
-                <Card.Text>{items.description}</Card.Text>
-                <hr />
-                <Card.Text>{items.address}</Card.Text>
-                <hr />
-                <Card.Text>Price: £{items.price} per month</Card.Text>
-                <hr />
-                {/*view button to display property data in modal*/}
-                <div className="d-flex flex-md-wrap justify-content-around gap-2 mt-3">
-                  {button(
-                    () => {
-                      // change modal header
-                      dispatch({
-                        type: "CHANGE HEADER",
-                        payload: `Property Address: ${items.address}`,
-                      });
-                      // once clicked show modal
-                      dispatch({
-                        type: "CHANGE MODAL STATE",
-                        payload: true,
-                      });
-                      // load USERS DETAILS form
-                      dispatch({
-                        type: "WHICH FORM TO USE",
-                        payload: "READ PROPERTY",
-                      });
-                      // pushed data to view an object in form state
-                      dispatch({ type: "FORM DETAILS", payload: items });
-                    },
-                    "View",
-                    "outline-info",
-                  )}
-                </div>
-              </Card.Body>
-            </Card>
+      {data ? (
+        id === undefined ? (
+          data.map((items) => (
+            <DisplayPropertyCard items={items} key={items.id} />
           ))
-        : spinner()}
+        ) : (
+          <DisplayPropertyCard items={data} />
+        )
+      ) : (
+        spinner()
+      )}
     </div>
   );
 };
