--- frontend/src/components/pages/dashboards/UsersDashboard.jsx
@@ -1,9 +1,0 @@
-const UserDashboard = () => {
-  return (
-    <div>
-      <h1>User Dashboard</h1>
-    </div>
-  );
-};
-
-export default UserDashboard;
